install: support old ln versions without the -r option
Oldish enterprise-class Linux distributions have outdated versions of
coreutils whose ln command do not support the -r option. Also non-Linux
systems like FreeBSD don't support that option. Use a shell script that
mimics the missing functionality.

The script creates minimalist relative paths, like ln -r does, but in
order to simplify the logic, it requires that the arguments be absolute
paths and do not end with '/'. This is enough for our purposes.

Add configuration logic to detect if ln supports the -r option, based on
the logic used by Autoconf to check if ln supports the -s option.

Signed-off-by: Carlos Santos <unixmania@gmail.com>