Update Generic-PDF_Printer-PDF.ppd
Hi,

Here is a patch  for Generic PDF PPD to enable color/gray scale printing switching.  Original Generic PPD has no color switch, therefore users cannot make black and white printing with color printers.

It adds a PJL RENDERMODE command to PDF print job.  When choosing "Black and White", the printout will be in gray scale (rendered in gray scale by printers), even if PDF is made as color document.
Of course it cannot print in color if the original document is made in black and white or the printer is not color capable :)

I checked it worked with Chrome Browser on Ubuntu as well as on ChromeOS.

Regards,