Build system: Install individual filter executables also with universal filter
If we use the "universal" CUPS filter which runs any needed filter
chain for CUPS out of a single executable, we also install the
executables to call the individual filter function, so that PPD files
specifying filter executables (of cups-filters) in their
"*cupsFilter(2):" lines will continue to work.

Also updated the wording of the help messages for universal/individual
filter executables use in the ./configure script.