libcupsfilters: Do actual commit for the previous commit message
Normally, GIT does not accept the command "git commit" without
specifying files or "-a" for all modified files. Strangely enough, it
accepts it when files got deleted and then it only commits the
deletions. So the previous commit (f89c3136cb8) only contains the
deletions of the cupsfilters/image-sgi.h and
cupsfilters/image-sgilib.c and not the actual changes.

Now here we go with the changes.