Update cfCatalogLoad() calls for API change
In libcupsfilters we have added language/translation support to the `cfCatalog...()` API functions via https://github.com/OpenPrinting/libcupsfilters/pull/2. This changes the `cfCatalogLoad()` calls in libppd (both in the PPD generator for driverless printing). This commit updates them.

For a quick solution we supply NULL as language for now, resembling the old behavior. We look into language support in the PPD generator later.