Transfer CUPS' cupstestppd utility to ppdTest() library function
The only part of the PPD file support in CUPS was the cupstestppd
utility. It is a valuable piece of code to debug issues with PPD
files, and also to prevent broken PPD files from being used.

We do not simply transfer the command line utility, but instead,
convert it into a library function. This way and application which
handles PPD files can easily check the integrity of a user-supplied
PPD file.

First use case is the PostScript Printer Application which allows the
user to upload a PPD file via the web admin interface, for printers
for which there is no PPD file already included in the package. To
assure that the printer actually works with the user's PPD we will now
be able to check the PPD file by calling the ppdTest() library
function and warn the user in case of failure.

See https://github.com/OpenPrinting/ps-printer-app/issues/12