ppdTest(): Usage instructions for testppdfile should not be in library function
ppdTest() is a library function to be called by any program, not only
by the command line tool testppdfile, so it should not provide the
usage instructions for the command line tool.

The usage() function is now in the ppd/testppdfile.c file and is called
directly by the testppdfile utility.