ppd-filter.c: Generate media-col if media-col-default exists
We would generate `media-col` attribute only if we had `media-col-default`
attribute and `PageSize` or `media` options, which doesn't seem to be always
the case for `PageSize` and `media`. These options usually are not
default options and if the application doesn't send them, they are not
sent to filters.

The current implementation breaks image printing for (at least) printers
installed with Postscript driver - the printer outputs an empty sheet of
paper. It happens because `cfFilterImageToPDF()` doesn't get page size
by other means.