ppdLoadAttributes(): Find default page size also by dimensions
When converting a PPD file into printer IPP attributes the list of
page sizes from the PWG mapping cache is used, and here duplicate
sizes (different names but same size and same borderless state,
borderless or not borderless) are removed. If the default is the
second or a later mention of its size, the name is removed in the
cache and so the entry for the default not found, making us resort to
A4/Letter or the first in the list.

This caused failures with PPDs for Dymo label printers, as they have
duplicate page size entries for some sizes as labels of this size are
common with several different names and these they want to appear in
the menus of print dialogs. the default page size in these PPDs was
often the second mention of its size and so we get said failure.

The problem got discovered by the dymo-cups-drivers Debian/Ubuntu
package failing its autopkgtests on Ubuntu 23.04 which is the first
with cups-filters 2.x, and so with libppd.

Now, if the default size of the PPD cannot be found in the cache by
name, it is first checked for an entry with the same size, before we
fallback to A4/Letter or to the first entry.