PPD file generator: Set default color mode when printer attrs say "auto"
With the printer reporting "print-color-mode-default=auto" in its IPP
attributes "RGB" got never set as default when for the ColorModel
option when a color choice got found for "print-color-mode", but
"Gray" was set when there was a standard monochromr choice, leadin to
color printers getting grayscale as default color mode.

Now this is fixed and always the "best" color mode chosen as default.

See https://bugs.launchpad.net/bugs/2014976