All PPD files with "MirrorPrint" option cuased mirrored printout
If a PPD contains an option "MirrorPrint", the ppdFilterLoadPPD()
sent the option "mirror=true" to the filter functions, regardless
of the actual setting of "MirrorPrint" (which is usually "False"
by default).

This made all printouts on printers using a PPD file with a
"MirrorPrint" option coming out mirrored, with no way to counteract
this by any options or attributes.

This is fixed now by checking the setting of "MirrorPrint".

See https://bugs.launchpad.net/bugs/2018538