Fixed resolution handling when converting PPDs to printer IPP attributes
When converting the PPD file into printer IPP attributes, usually for
filter functions of libcupsfilters to be used in classic CUPS filters,
the resolution was not handled correctly.

If the intended output is PWG Raster, Apple Raster, or PCLm,
specifying a resolution in the job IPP attributes or the CUPS job
options is ignored, also a default resolution specified in the PPD
file or a resolution resulting from pseudo-PostScript assigned to the
option choices in the PPD. Instead, always the first (lowest)
resolution in the list of supported resolutions for the output Raster
format is used.

Now we respect job attributes, pseudo-PostScript code, and PPD's
default resolution with priority, and fall back to the base resolution
for the output Raster format only if everything else fails.

In many cases we get higher output resolution by that.