ppdFilterLoadPPD(): Actually create Raster header also for Apple/PWG Raster
Not only if the PPD's "*cupsFilter(2): ..." attributes specify CUPS
Raster as format needed from the CUPS filters but also for Apple
Raster or PWG Raster a sample header for the filter functions should
be generated.

This commit fixes the additional checks for the "cupsUrfSupported" and
"cupsPwgRasterDocumentTypeSupported" PPD attributes and now the sample
header gets created for all Raster formats.