Added support for libcups3 (libcups of CUPS 3.x) (#27)
With these changes libppd can be built either with libcups2
(libcups of CUPS 2.x) or libcups3 (libcups of CUPS 3.x).

    NOTE: Any further development on the code has to be done using the
    API of libcups3, and for libcups2 support any additional
    conversion has to get defined in cupsfilters/libcups2-private.h
    (if C code is needed for the conversion, it goes into
    cupsfilters/libcups2.c).

The actual code is now fully converted to the API of libcups3. To
allow building it with libcups2 and to avoid having it full lot of "#ifdef
HAVE_LIBCUPS2" conditionals, we have added the header file
libcups2-private.h file, which gets only applied (by "#ifdef
HAVE_LIBCUPS2") when we build with libcups2.

It converts the code to the libcups2 API by

- Applying "#define ..." macro definitions for all renamings of
  functions, data types, and constants

- Converting the "cupsCreateTempFd(NULL, NULL, buffer, bufsize)" calls
  to "cupsTempFd(buffer, bufsize)" calls (new function has more
  arguments).

- Converting the "cupsCreateTempFile(NULL, NULL, buffer, bufsize)" calls
  to "cupsTempFile2(buffer, bufsize)" calls (new function has more
  arguments).

- Introducing the type "cups_len_t", as "int" for libcups2 and
  "size_t" for libcups3. In some functions there are parameters which
  make a hard switch from "int" to "size_t" with the transition to
  libcups3. For those we use this type.

- Also introducing "cups_bool_t" as "int" for libcups2 and "bool" for
  libcups3, for functions where parameters which were of type "int"
  got passed to "bool".

We also add the file libcups2.c for implementing the functions
"cupsLangGetName()" and "cupsLangGetString()", newly introduced
in libcups3 as a version for libcups2.

The files ppd/language.c and ppd/language-private.h got removed and
their functionality passed into ppd/ppd-cache.c and ppd/libcups2.c, as
there are no other files in the library using it.

In ppd/ppd-generator.c introduced data type "_ppd_size_t" to replace
the identical "cups_size_t" type which only exists in libcups2, this way
we can have the code identical for both libcups versions.

At some places we simply use casts to "(int)"instead of using the
"cups_len_t" an "cups_bool_t" types.

In ppd/ppdc-catalog.cxx we have calls of "cupsFileWrite()" which returned
the written number of bytes as "int" value. In libcups3 the function
returns success or failure as "bool". Here we had no other solution than
using "#ifdef HAVE_LIBCUPS2" conditionals.

We have a libcups3-related change in the public API header files, here
"cups_page_header2_t" got renamed into "cups_page_header_t". Due to
the fact that libcups2-private.h gets HAVE_LIBCUPS2 from config.h, we
cannot use it in the API headers (and we also do not want all these
conversions to apply to a user's code). So we simply add a renaming
"#define ..." under "#if CUPS_VERSION_MAJOR < 3" to the affected API
headers.

In configure.ac we check for libcups3 via pkg-config first and if not
present for libcups2 via cups-config. In both cases we set all relevant
variables, as compiler/linker arguments, directories, ... and in case
of libcups2 we define HAVE_LIBCUPS2 in config.h. We also do
corrections on the CXXFLAGS, adding "-std=c++17" and "-Werror"
(if the user opts for warnings treated as errors).

In Makefile.am we do not only add the new files to the build process
but we also add the compiler flag definition for C++,
"libppd_la_CXXFLAGS", simply being a copy of "libppd_la_CFLAGS".
Without it, the header files of libcups3 are not found when compiling
C++ code. We also do corrections for the build rules of the build test program
ppd/testppd.c