ppd-cache.c: Use PPD size name from CUPS if size is standardized
Previously, if there were two sizes with the same size, but different
names in PPD, the second was ignored during generating sizes for PPD
cache. This behavior breaks already broken PPDs, which provide two or
more different size names for the same width and length (which is
incorrect), and if the ignored size was the PPD default (such as A4),
user couldn't print a test page.

After this fix, we use PPD size name from CUPS if the size is
standarized.

Fixes #29