ppdCreatePPDFromIPP2(): Also check for "pwg-raster-document-type-supported"
If the printer supports printing input in PWG Raster format, it should
not only tell that it accepts PWG Raster in
"document-format-supported" but also provide the attributes with the
specifications of what raster formats are exactly supported, and here
not only resolutions via "pwg-raster-document-resolution-supported"
but also color spaces and depths via
"pwg-raster-document-type-supported". If one of these two attributes
is missing, we disconsider PWG Raster.

See also the discussion in

https://github.com/OpenPrinting/cups/pull/890

Thanks Zdenek Dohnal and Michael Sweet.